working on adding bigInteger to project

+++ Smart Calculator/task/src/calculator/Main.java
@@ -2,10 +2,11 @@ package calculator;
 
 import java.util.HashMap;
 import java.util.Scanner;
+import java.math.BigInteger;
 
 public class Main {
 
-    public static HashMap<String, Integer> variables;
+    public static HashMap<String, BigInteger> variables;
 
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
